Move unecessary Dimension.integ.spec to Dimension.unit.spec and fix broken tests
